==> notebooks/01_data_loading.ipynb <==
from libs.load_datasets import load_csv_files

## 1. Loading all the Data sets

# Exemple d'utilisation de la fonction
dfs = load_csv_files("../data/inputs")

print(dfs['circuits'].shape)
dfs['circuits'].head()

print(dfs['status'].shape)
dfs['status'].head(15)

dfs['lap_times'].head()


dfs['sprint_results'].head()


dfs['drivers'].head()


dfs['races'].head()


print(dfs['constructors'].shape)
dfs['constructors'].head()

dfs['constructors'][dfs['constructors'].name=='Alpine']


dfs['constructor_standings'].head()


dfs['qualifying'].head()


print(dfs['driver_standings'].shape)
dfs['driver_standings'].head()


print(dfs['constructor_results'].shape)
dfs['constructor_results'].tail(10)
# dfs['status'][dfs['status'].notna()].tail(10)


dfs['pit_stops'].head()


dfs['seasons'].head()


print(dfs['results'].shape)
dfs['results'].head()
