==> notebooks/02_race_first_analysis.ipynb <==
%matplotlib inline
import pandas as pd
import plotly.express as px
import plotly.io as pio

pio.templates.default = "plotly_white"

## 1. Loading races Data set

df_races = pd.read_csv('../data/inputs/races.csv')
df_circuits = pd.read_csv('../data/inputs/circuits.csv')

df_races.head()

df_circuits.head()

df_races.sort_values(by='year', ascending=False).head()

df_race_v1 = df_races.merge(df_circuits[['circuitId', 'country']], how='inner', on='circuitId')

## 2. Race Analysis

### 2.1 Amount of races by year

px.histogram(df_race_v1, x='year', y='raceId', histfunc='count', nbins=90, range_x=[1950, 2023])

### 2.2 Amount global of races by circuit

px.histogram(df_race_v1, y='raceId', x='country', histfunc='count')
