==> notebooks/04_result_processing.ipynb <==
%matplotlib inline
import pandas as pd

from libs.time_transformer import time_to_milliseconds

## 1. Loading the Data set

df_results = pd.read_csv('../data/inputs/results.csv')

print(df_results.shape)
df_results.head()

## 2. Converting fatestLapTime in milliseconds

### 2.1. add a default values to a results without fastestLapTime defined (\N)

# Filter rows where fastestLapTime is '\\N' and replace with '90:00'
df_results.loc[df_results.fastestLapTime == '\\N', 'fastestLapTime'] = '90:00'

### 2.2. Convert a fastestLapTime field

# Convert 'fastestLapTime' to milliseconds using vectorized operation
df_results['fastestLapTime'] = df_results['fastestLapTime'].apply(time_to_milliseconds)

df_results.head()

## 3. Delete the time field and add a default to \N in time

df_results.drop('time', axis=1, inplace=True)

df_results.loc[df_results.milliseconds == '\\N', 'milliseconds'] = '0'


df_results.tail(10)

## 4. Remove unnecessary fields

Because we can have these informations in position Order, for the number it is present in the driver dataset

df_results.drop(['positionText', 'number', 'position', 'rank'], axis=1, inplace=True)

df_results.head()

## 5. Put the right type to each fields

df_results.info()

### 5.1. Fill the \N value (NA) in the fields that has type object

df_results.loc[df_results.milliseconds == '\\N']

df_results.loc[df_results.fastestLap == '\\N', 'fastestLap'] = 0
df_results.loc[df_results.fastestLap == '\\N']

df_results.loc[df_results.fastestLapSpeed == '\\N', 'fastestLapSpeed'] = 0
df_results.loc[df_results.fastestLapSpeed == '\\N']

### 5.1. Int fields to int64

df_results.milliseconds = df_results.milliseconds.astype('int64')
df_results.fastestLap = df_results.fastestLap.astype('int64')
df_results.fastestLapSpeed = df_results.fastestLapSpeed.astype('float64')

df_results.info()

df_results.to_csv('../data/outputs/01_results_cleaned.csv', index=False)
