==> notebooks/05_merging_results_and_race.ipynb <==
%matplotlib inline
import pandas as pd

## 1. Loading the Data set

df_races = pd.read_csv('../data/inputs/races.csv')
df_results = pd.read_csv('../data/outputs/01_results_cleaned.csv')

print(df_races.shape)
df_races.head()

print(df_results.shape)
df_results.head()

## 2. Merging the two Data set

df_race_results = df_results.merge(df_races[['raceId', 'year', 'round', 'circuitId', 'name']], how='inner', on='raceId')

print(df_race_results.shape)

print(df_race_results.shape)
df_race_results.head()

df_race_results.to_csv("../data/outputs/02_races_results.csv", index=False)
