==> notebooks/06_driver_processing.ipynb <==
Processing to do
---
- Nombre moyen de points par saison puit retirer la colonne point
- Vitesse de pointe moyenne par saison 
- Nombre de courses terminee par saison
- le circuit ayant le Nombre max de victoire 
- compter le nombre de fois premier
- compter le nombre de fois dans le top 10

%matplotlib inline
import pandas as pd
import numpy as np

## 1. Loading the Data sets

df_drivers = pd.read_csv("../data/inputs/drivers.csv")
df_races_results = pd.read_csv("../data/outputs/02_races_results.csv")

df_drivers.drop(['driverRef', 'url'], axis=1, inplace=True)

df_drivers.head()

df_races_results.head()

## 2. Add a flag for actives drivers

df_drivers['driver_is_active'] = 0

list_drivers = [
    ("Max", "Verstappen"),
    ("Logan", "Sargeant"),
    ("Lando", "Norris"),
    ("Pierre", "Gasly"),
    ("Sergio", "Pérez"),
    ("Fernando", "Alonso"),
    ("Charles", "Leclerc"),
    ("Lance", "Stroll"),
    ("Kevin", "Magnussen"),
    ("Yuki", "Tsunoda"),
    ("Alexander", "Albon"),
    ("Esteban", "Ocon"),
    ("Lewis", "Hamilton"),
    ("Carlos", "Sainz"),
    ("George", "Russell"),
    ("Oscar", "Piastri"),
    ("Guanyu", "Zhou"),
    ("Nico", "Hülkenberg"),
    ("Valtteri", "Bottas")
]

df_drivers.head()

for (name, surname) in list_drivers:
    df_drivers.loc[(df_drivers['forename'] == name) & (df_drivers['surname'] == surname), 'driver_is_active'] = 1

df_drivers[df_drivers.driver_is_active == 1].head()

## 3. Add the calculated features

### 3.1. Average of points by driver

ids = df_drivers.driverId.to_list()

df_drivers['driver_avg_point'] = np.float64(0)

for driverId in ids:
    avg_points = df_races_results[df_races_results.driverId == driverId]['points'].mean()
    df_drivers.loc[df_drivers.driverId == driverId, 'driver_avg_point'] = np.float64(avg_points)

df_drivers.head()

### 3.2. Average speed by driver

df_drivers['driver_avg_speed'] = np.float64(0)

for driverId in ids:
    avg_speed = df_races_results[(df_races_results.driverId == driverId) & (df_races_results.fastestLapSpeed > 0)]['fastestLapSpeed'].mean()
    df_drivers.loc[df_drivers.driverId == driverId, 'driver_avg_speed'] = np.float64(avg_speed)

df_drivers.fillna({'driver_avg_speed': 0}, inplace=True)

# Find the avg speed of Max Verstappen
df_drivers[df_drivers.forename.str.contains('Max')].head()

### 3.3 Number of races ends by seaoson (From 2015)

years_list = [0, 2019, 2020, 2021, 2022, 2023]
df_drivers['race_end_bf_2019'] = 0
df_drivers['race_end_in_2019'] = 0
df_drivers['race_end_in_2020'] = 0
df_drivers['race_end_in_2021'] = 0
df_drivers['race_end_in_2022'] = 0
df_drivers['race_end_in_2023'] = 0

for driverId in ids:
    for year in years_list:
        if 0 == year:
            nber_of_win = df_races_results[(df_races_results.driverId == driverId) & (df_races_results.statusId == 1) & (df_races_results.year < 2019)].raceId.count()
            df_drivers.loc[df_drivers.driverId == driverId, 'race_end_bf_2019'] = nber_of_win
        else:
            nber_of_win = df_races_results[(df_races_results.driverId == driverId) & (df_races_results.statusId == 1) & (df_races_results.year == year)].raceId.count()
            df_drivers.loc[df_drivers.driverId == driverId, f'race_end_in_{year}'] = nber_of_win

df_drivers.head()

### 3.4. Most won circuits

df_drivers['driver_most_won_circuit_id'] = 0

for driverId in ids:
    try:
        df_drivers.loc[df_drivers.driverId == driverId, 'driver_most_won_circuit_id'] = df_races_results[
            (df_races_results.driverId == driverId) & (df_races_results.positionOrder == 1)
                ].groupby(['circuitId', 'name'])[['resultId']].count().sort_values(
                    ascending=False, by='resultId').resultId.idxmax()[0]
    except:
        df_drivers.loc[df_drivers.driverId == driverId, 'driver_most_won_circuit_id'] = 0

df_drivers.head()

### 3.5. Number of races won

df_drivers['driver_nber_of_races_won'] = 0

for driverId in ids:
    df_drivers.loc[df_drivers.driverId == driverId, 'driver_nber_of_races_won'] = df_races_results[(df_races_results.driverId == driverId) & (df_races_results.positionOrder == 1)].shape[0]

df_drivers.head()

### 3.6. Number of times in the top 10 table

df_drivers['driver_nber_of_times_in_top_10'] = 0

for driverId in ids:
    df_drivers.loc[df_drivers.driverId == driverId, 'driver_nber_of_times_in_top_10'] = df_races_results[(df_races_results.driverId == driverId) & (df_races_results.positionOrder <= 10)].shape[0]

df_drivers.head()

df_drivers['dob'] = pd.to_datetime(df_drivers['dob'])

df_drivers['age'] = pd.to_datetime('now').year - df_drivers['dob'].dt.year

df_drivers.to_csv('../data/outputs/03_drivers_processed.csv', index=False)
